#159947782 Crud for Recipes (#4)
* [Feature #159947782]Add create recipe functionality

This commit enables an authenticated user to create a recipe provided they have provided a category id in the url. This category has to be a ctegory created by them.

* [feature #159947782]Add CRUD for recipes

* [Feature #159947782] Add CRUD functionality for recipes

This commit enables a user to create,list,update,retrieve and delete recipes. This commit also integrate TravisCI to setup the CI pipeline.

* [Fix] declare requirements.txt