Develop (#6)
* [Chore #159947745] Initail project setup (#1)

This commit sets up the initail project. It caters for databse connection to a postgres db, installing necessary dependacies and url files for the API

* [Feature #159947753] Add User registration (#2)

This commit adds user registration capability . A User needs to add their email,username and password.

* [Feature #160163254]Add Category CRUD (#3)

This commit caters to creation,reading,updating and deletion of categories.This commit also introduces token authentication and a user needs to provide a token in the headers inorder to carry out any of the CRUD fuctionality for categories

* #159947782 Crud for Recipes (#4)

* [Feature #159947782]Add create recipe functionality

This commit enables an authenticated user to create a recipe provided they have provided a category id in the url. This category has to be a ctegory created by them.

* [feature #159947782]Add CRUD for recipes

* [Feature #159947782] Add CRUD functionality for recipes

This commit enables a user to create,list,update,retrieve and delete recipes. This commit also integrate TravisCI to setup the CI pipeline.

* [Fix] declare requirements.txt

* [Chore] Add read me

* [chore] Add Badge